remove previous hack to always place doc in master
Back in EE-1.7 I was having troubles placing the .rst into the
dev/ee-1.7 branch using the script.  I ended up just pushing the
doc into master.
Remove the hack such that the branch for which we store the doc
is the same branch from which we obtained the trove-client.

Change-Id: I3bc0e395b3091e3beb3e4d86575669c112507e29